# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Omersia Storefront API",
        "description": "Documentation de l'API publique Storefront pour l'e-commerce Omersia",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Environnement local / base"
        }
    ],
    "paths": {
...
        "/api/v1/cart/sync": {
            "post": {
                "tags": [
                    "Panier"
                ],
                "summary": "Synchroniser le panier (créer ou mettre à jour)",
                "operationId": "e6f2c4a53923b74252c5a7d379061a94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartSyncRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Panier synchronisé ou supprimé. Si items vide/null, retourne {deleted: true}. Sinon retourne le panier mis à jour.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "description": "Panier mis à jour",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "cart_abc123"
                                                },
                                                "subtotal": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 59.98
                                                },
                                                "total_qty": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "EUR"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "description": "Panier supprimé (items vides)",
                                            "properties": {
                                                "deleted": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "API key invalide ou manquante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api.key": []
                    }
                ]
            }
        },
        "/api/v1/cart/apply-discount": {
            "post": {
                "tags": [
                    "Panier"
                ],
                "summary": "Appliquer un code promo au panier",
                "operationId": "6e3d8069ae6c75a65aaba7ed99de2a25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DiscountApplicationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code promo appliqué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountApplicationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "API key invalide ou manquante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Code promo invalide ou non applicable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code promo invalide ou expiré."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api.key": []
                    }
                ]
            }
        },
        "/api/v1/cart/apply-automatic-discounts": {
            "post": {
                "tags": [
                    "Panier"
                ],
                "summary": "Appliquer automatiquement les réductions actives au panier",
                "operationId": "2575a11a101a0f5588fda8ef67d7695c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Customer ID",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "qty",
                                                "price"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "description": "Product ID",
                                                    "type": "integer"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "minimum": 0
                                                },
                                                "variant_id": {
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réductions automatiques évaluées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "promotions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "type": "string",
                                                        "example": "AUTO-12"
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "Soldes été"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "order"
                                                    },
                                                    "value_type": {
                                                        "type": "string",
                                                        "example": "percentage"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "line_adjustments_by_code": {
                                            "description": "Ajustements ligne par code promo",
                                            "type": "object"
                                        },
                                        "order_discount_total": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 10
                                        },
                                        "product_discount_total": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5
                                        },
                                        "shipping_discount_total": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0
                                        },
                                        "free_shipping": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "API key invalide ou manquante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api.key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartSyncRequest": {
                "description": "Cart synchronization request",
                "properties": {
                    "token": {
                        "description": "Cart token for guest users",
                        "type": "string",
                        "example": "cart_abc123xyz",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "customer@example.com",
                        "nullable": true
                    },
                    "currency": {
                        "description": "Currency code (3 chars)",
                        "type": "string",
                        "example": "EUR",
                        "nullable": true
                    },
                    "items": {
                        "description": "Cart items. If empty or null, the cart will be deleted.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DiscountApplicationRequest": {
                "description": "Apply discount code request",
                "required": [
                    "code",
                    "items"
                ],
                "properties": {
                    "code": {
                        "description": "Discount code",
                        "type": "string",
                        "example": "SUMMER2024"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "customer@example.com",
                        "nullable": true
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 59.98
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        }
                    }
                },
                "type": "object"
            },
            "DiscountApplicationResponse": {
                "description": "Discount application result",
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "code": {
                        "description": "Code promo appliqué",
                        "type": "string",
                        "example": "SUMMER2024"
                    },
                    "label": {
                        "description": "Nom de la réduction",
                        "type": "string",
                        "example": "Soldes été 2024"
                    },
                    "type": {
                        "description": "Type de réduction",
                        "type": "string",
                        "enum": [
                            "order",
                            "product",
                            "shipping",
                            "buy_x_get_y"
                        ],
                        "example": "order"
                    },
                    "value_type": {
                        "description": "Type de valeur",
                        "type": "string",
                        "enum": [
                            "percentage",
                            "fixed"
                        ],
                        "example": "percentage"
                    },
                    "value": {
                        "description": "Valeur de la réduction",
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "discount_amount": {
                        "description": "Montant total de la réduction",
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "order_discount_amount": {
                        "description": "Montant réduction sur la commande",
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "product_discount_amount": {
                        "description": "Montant réduction sur les produits",
                        "type": "number",
                        "format": "float",
                        "example": 5
                    },
                    "shipping_discount_amount": {
                        "description": "Montant réduction sur la livraison",
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "free_shipping": {
                        "description": "Livraison gratuite accordée",
                        "type": "boolean",
                        "example": false
                    },
                    "line_adjustments": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Product ID",
                                    "type": "integer"
                                },
                                "variant_id": {
                                    "description": "Variant ID",
                                    "type": "integer",
                                    "nullable": true
                                },
                                "discount_amount": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "is_gift": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "description": "Validation error response (HTTP 422)",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "description": "Generic validation error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Field-specific validation errors",
                        "type": "object",
                        "example": {
                            "email": [
                                "L'adresse email est déjà utilisée."
                            ],
                            "password": [
                                "Le mot de passe doit contenir au moins 8 caractères."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "ApiKeyError": {
                "description": "API key validation error",
                "properties": {
                    "message": {
                        "description": "API key error message",
                        "type": "string",
                        "example": "Invalid API key"
                    },
                    "error": {
                        "description": "Detailed error (e.g., expired, missing)",
                        "type": "string",
                        "example": "API key expired",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ServerError": {
                "description": "Internal server error (HTTP 500)",
                "properties": {
                    "message": {
                        "description": "Generic server error message",
                        "type": "string",
                        "example": "Server Error"
                    },
                    "error": {
                        "description": "Detailed error (only in debug mode)",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api.key": {
                "type": "apiKey",
                "description": "Frontend API Key (required for all requests)",
                "name": "X-API-KEY",
                "in": "header"
            },
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
    "security": [
        {
            "api.key": []
        }
    ]
}